logging: Break out logging functionality into separate file
